refactor(engine): move inline tests to engine/tests/ integration dir
Extracted all #[cfg(test)] mod tests { ... } blocks from engine/src/ into
dedicated files under engine/tests/, creating a new cli module for CLI helpers
that were previously only accessible from the binary. All 246 tests pass.
